First Publish
Code runs through different actions (open gripper, close gripper, move to place 45, move to place 25... etc) in an order to calibrate values such as the motor count at the different positions. The bluemotor.cpp code was updated to be non blocking. The main.cpp state machine is non blocking and is controlled by the IR sensor.